pyproject.toml
[project]
name = "odoo-bringout-stock_mto_purchase_with_analytic"
version = "16.0.0"
description = "Stock MTO Purchase with Analytic - Adds analytic distribution to purchase orders generated from manufacturing orders."
authors = [
    { name = "Ernad Husremovic", email = "hernad@bring.out.ba" }
]
dependencies = [
    "odoo-bringout-mrp_stock_analytic>=16.0.0",
    "odoo-bringout-stock_mts_mto_rule>=16.0.0",
    "odoo-bringout-purchase_analytic>=16.0.0",
    "odoo-bringout-purchase_stock_analytic>=16.0.0",
    "requests>=2.25.1"
]
readme = "README.md"
requires-python = ">= 3.11"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11", 
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]

[project.urls]
homepage = "https://github.com/bringout/0"
repository = "https://github.com/bringout/0"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["stock_mto_purchase_with_analytic"]

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.4.1",
]
